refactor: rename authentication views
